upgrade inject to 0.4.1
